share versions with meta-build, update clue
