+++ project/Versions.scala
@@ -1,0 +1,44 @@
+object Versions {
+  val boopickle              = "1.4.0"
+  val cats                   = "2.9.0"
+  val catsEffect             = "3.5.0-RC3"
+  val catsRetry              = "3.1.0"
+  val catsTime               = "0.5.1"
+  val circe                  = "0.14.5"
+  val circeGolden            = "0.3.0"
+  val coulomb                = "0.7.3"
+  val clue                   = "0.30.0"
+  val crystal                = "0.33.11"
+  val discipline             = "1.5.1"
+  val disciplineMUnit        = "1.0.9"
+  val fs2                    = "3.6.1"
+  val fs2Data                = "1.7.0"
+  val fs2Dom                 = "0.1.0"
+  val geminiLocales          = "0.7.0"
+  val http4s                 = "0.23.18"
+  val http4sDom              = "0.2.7"
+  val kittens                = "3.0.0"
+  val log4Cats               = "2.5.0"
+  val log4CatsLogLevel       = "0.3.1"
+  val lucumaBC               = "0.4.0"
+  val lucumaCore             = "0.74.0"
+  val lucumaCatalog          = "0.40.3"
+  val lucumaReact            = "0.33.0"
+  val lucumaRefined          = "0.1.1"
+  val lucumaSchemas          = "0.49.0"
+  val lucumaSSO              = "0.5.7"
+  val lucumaUI               = "0.69.1"
+  val monocle                = "3.2.0"
+  val mouse                  = "1.2.1"
+  val mUnit                  = "0.7.29"
+  val mUnitCatsEffect        = "1.0.7"
+  val pprint                 = "0.8.1"
+  val reactAladin            = "0.29.1"
+  val refinedAlgebra         = "0.1.0"
+  val sbtBuildInfo           = "0.11.0"
+  val sbtLucuma              = "0.10.12"
+  val scalaCollectionContrib = "0.3.0"
+  val scalaJsDom             = "2.4.0"
+  val scalaJsReact           = "2.1.1"
+  val webAppUtil             = "2.0.0-RC12"
+}
