+++ project/project/Versions.scala
@@ -1,0 +1,1 @@
+../Versions.scala
